Move json formatter to utils Move parser to utils
